organize dashapp a little better
